Create tabs for different charts.
